+++ src/main/java/lk/recruitment_management/asset/applicant/controller/ApplicantController.java
@@ -4,8 +4,16 @@ package lk.recruitment_management.asset.applicant.controller;
 import lk.recruitment_management.asset.ag_office.controller.AgOfficeController;
 import lk.recruitment_management.asset.ag_office.service.AgOfficeService;
 import lk.recruitment_management.asset.applicant.entity.*;
-import lk.recruitment_management.asset.applicant.entity.Enum.*;
-import lk.recruitment_management.asset.applicant.service.ApplicantFilesService;
+import lk.recruitment_management.asset.applicant.entity.enums.*;
+import lk.recruitment_management.asset.applicant_file.service.ApplicantFilesService;
+import lk.recruitment_management.asset.applicant_degree_result.entity.ApplicantDegreeResult;
+import lk.recruitment_management.asset.applicant_file.entity.ApplicantFiles;
+import lk.recruitment_management.asset.applicant_gazette.entity.enums.ApplyingRank;
+import lk.recruitment_management.asset.applicant_result.entity.ApplicantResult;
+import lk.recruitment_management.asset.applicant_result.entity.enums.Attempt;
+import lk.recruitment_management.asset.applicant_result.entity.enums.CompulsoryOLSubject;
+import lk.recruitment_management.asset.applicant_result.entity.enums.StreamLevel;
+import lk.recruitment_management.asset.applicant_result.entity.enums.SubjectResult;
 import lk.recruitment_management.asset.common_asset.model.Enum.*;
 import lk.recruitment_management.asset.common_asset.model.TwoDate;
 import lk.recruitment_management.asset.common_asset.service.CommonService;
@@ -14,6 +22,7 @@ import lk.recruitment_management.asset.district.controller.DistrictController;
 import lk.recruitment_management.asset.district.service.DistrictService;
 import lk.recruitment_management.asset.grama_niladhari.controller.GramaNiladhariController;
 import lk.recruitment_management.asset.grama_niladhari.service.GramaNiladhariService;
+import lk.recruitment_management.asset.non_relative.entity.NonRelative;
 import lk.recruitment_management.asset.police_station.controller.PoliceStationController;
 import lk.recruitment_management.asset.police_station.Service.PoliceStationService;
 import lk.recruitment_management.asset.user_management.service.UserService;
