--- src/main/java/lk/recruitment_management/asset/applicant/entity/ApplicantFiles.java
@@ -1,41 +1,0 @@
-package lk.recruitment_management.asset.applicant.entity;
-
-
-import lk.recruitment_management.util.audit.AuditEntity;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Lob;
-import javax.persistence.ManyToOne;
-
-@Entity
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class ApplicantFiles extends AuditEntity {
-
-    private String name, mimeType,newName;
-
-    @Column(unique = true)
-    private String newId;
-
-    @Lob
-    private byte[] pic;
-
-    public ApplicantFiles(String name, String mimeType, byte[] pic, String newName, String newId) {
-        this.name = name;
-        this.mimeType = mimeType;
-        this.pic = pic;
-        this.newName = newName;
-        this.newId = newId;
-    }
-
-    @ManyToOne
-    private Applicant applicant;
-
-}
