--- src/main/java/lk/recruitment_management/asset/applicant/entity/Enum/Attempt.java
@@ -1,13 +1,0 @@
-package lk.recruitment_management.asset.applicant.entity.Enum;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum Attempt {
-    FIRST("First"),
-    SECOND("Second");
-
-    private final String attempt;
-}
