--- src/main/java/lk/recruitment_management/asset/applicant/entity/Enum/CompulsoryOLSubject.java
@@ -1,17 +1,0 @@
-package lk.recruitment_management.asset.applicant.entity.Enum;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum CompulsoryOLSubject {
-    MATH("Mathematics"),
-    SCIEN("Science"),
-    SING("Sinhala Language & Literature"),
-    ENG("English Language & Literature"),
-    REG("Religion"),
-    OTH("Other");
-
-    private final String compulsoryOLSubject;
-}
