--- src/main/java/lk/recruitment_management/asset/applicant/entity/Enum/Nationality.java
@@ -1,17 +1,0 @@
-package lk.recruitment_management.asset.applicant.entity.Enum;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum Nationality {
-
-    SIN("Sinhala"),
-    TML("Tamil"),
-    MLM("Muslim"),
-    BGR("Burghers");
-
-
-    private final String nationality;
-}
