--- src/main/java/lk/recruitment_management/asset/applicant/entity/Enum/StreamLevel.java
@@ -1,13 +1,0 @@
-package lk.recruitment_management.asset.applicant.entity.Enum;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum StreamLevel {
-    OL("Ordinary Level"),
-    AL("Advance Level");
-
-    private final String streamLevel;
-}
