--- src/main/java/lk/recruitment_management/asset/applicant/entity/Enum/StudentStatus.java
@@ -1,16 +1,0 @@
-package lk.recruitment_management.asset.applicant.entity.Enum;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum StudentStatus {
-    ACTIVE("Active"),
-    BATCHDROP("Batch Drop"),
-    COURSEDROP("Course Drop"),
-    NOTCOMPLETE("Not Complete"),
-    COMPLETED("Completed");
-
-    private final String studentStatus;
-}
