--- src/main/java/lk/recruitment_management/asset/applicant/entity/Enum/SubjectResult.java
@@ -1,16 +1,0 @@
-package lk.recruitment_management.asset.applicant.entity.Enum;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum SubjectResult {
-    A("A"),
-    B("B"),
-    C("C"),
-    S("S"),
-    W("W");
-
-    private final String subjectResult;
-}
