+++ src/main/java/lk/recruitment_management/asset/applicant/entity/enums/Nationality.java
@@ -1,0 +1,17 @@
+package lk.recruitment_management.asset.applicant.entity.enums;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum Nationality {
+
+    SIN("Sinhala"),
+    TML("Tamil"),
+    MLM("Muslim"),
+    BGR("Burghers");
+
+
+    private final String nationality;
+}
