+++ src/main/java/lk/recruitment_management/asset/applicant/service/ApplicantService.java
@@ -8,12 +8,12 @@ import com.itextpdf.text.pdf.PdfWriter;
 import com.itextpdf.text.Font;
 import lk.recruitment_management.asset.applicant.dao.ApplicantDao;
 import lk.recruitment_management.asset.applicant.entity.Applicant;
-import lk.recruitment_management.asset.applicant.entity.ApplicantDegreeResult;
-import lk.recruitment_management.asset.applicant.entity.ApplicantResult;
-import lk.recruitment_management.asset.applicant.entity.Enum.ApplicantStatus;
-import lk.recruitment_management.asset.applicant.entity.Enum.ApplyingRank;
-import lk.recruitment_management.asset.applicant.entity.Enum.StreamLevel;
-import lk.recruitment_management.asset.applicant.entity.NonRelative;
+import lk.recruitment_management.asset.applicant_degree_result.entity.ApplicantDegreeResult;
+import lk.recruitment_management.asset.applicant_result.entity.ApplicantResult;
+import lk.recruitment_management.asset.applicant.entity.enums.ApplicantStatus;
+import lk.recruitment_management.asset.applicant_gazette.entity.enums.ApplyingRank;
+import lk.recruitment_management.asset.applicant_result.entity.enums.StreamLevel;
+import lk.recruitment_management.asset.non_relative.entity.NonRelative;
 import lk.recruitment_management.asset.interview.entity.Enum.InterviewName;
 import lk.recruitment_management.asset.interview.service.InterviewService;
 import lk.recruitment_management.asset.interview_parameter.entity.InterviewParameter;
