--- src/main/java/lk/recruitment_management/asset/applicant/service/ApplicantSubjectResultFilesService.java
@@ -1,4 +1,0 @@
-package lk.recruitment_management.asset.applicant.service;
-
-public class ApplicantSubjectResultFilesService {
-}
