+++ src/main/java/lk/recruitment_management/asset/applicant_gazette/entity/ApplicantGazette.java
@@ -3,6 +3,7 @@ package lk.recruitment_management.asset.applicant_gazette.entity;
 import com.fasterxml.jackson.annotation.JsonFilter;
 import lk.recruitment_management.asset.applicant.entity.Applicant;
 import lk.recruitment_management.asset.applicant_gazette.entity.enums.ApplicantGazetteAttempt;
+import lk.recruitment_management.asset.applicant_gazette.entity.enums.ApplyingRank;
 import lk.recruitment_management.asset.applicant_gazette_interview.entity.ApplicantGazetteInterview;
 import lk.recruitment_management.asset.gazette.entity.Gazette;
 import lk.recruitment_management.util.audit.AuditEntity;
@@ -30,6 +31,9 @@ public class ApplicantGazette extends AuditEntity {
 
   @ManyToOne
   private Gazette gazette;
+
+  @Enumerated(EnumType.STRING)
+  private ApplyingRank applyingRank;
 
   @OneToMany(mappedBy ="applicantGazette" )
   private List< ApplicantGazetteInterview > applicantGazetteInterviews;
