+++ src/main/java/lk/recruitment_management/asset/applicant_result/entity/enums/CompulsoryOLSubject.java
@@ -1,0 +1,17 @@
+package lk.recruitment_management.asset.applicant_result.entity.enums;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum CompulsoryOLSubject {
+    MATH("Mathematics"),
+    SCIEN("Science"),
+    SING("Sinhala Language & Literature"),
+    ENG("English Language & Literature"),
+    REG("Religion"),
+    OTH("Other");
+
+    private final String compulsoryOLSubject;
+}
