+++ src/main/java/lk/recruitment_management/asset/applicant_result/entity/enums/StreamLevel.java
@@ -1,0 +1,13 @@
+package lk.recruitment_management.asset.applicant_result.entity.enums;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum StreamLevel {
+    OL("Ordinary Level"),
+    AL("Advance Level");
+
+    private final String streamLevel;
+}
