+++ src/main/java/lk/recruitment_management/asset/applicant_result/entity/enums/SubjectResult.java
@@ -1,0 +1,16 @@
+package lk.recruitment_management.asset.applicant_result.entity.enums;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum SubjectResult {
+    A("A"),
+    B("B"),
+    C("C"),
+    S("S"),
+    W("W");
+
+    private final String subjectResult;
+}
