+++ src/main/java/lk/recruitment_management/asset/common_asset/model/TwoDate.java
@@ -1,7 +1,7 @@
 package lk.recruitment_management.asset.common_asset.model;
 
-import lk.recruitment_management.asset.applicant.entity.Enum.ApplicantStatus;
-import lk.recruitment_management.asset.applicant.entity.Enum.ApplyingRank;
+import lk.recruitment_management.asset.applicant.entity.enums.ApplicantStatus;
+import lk.recruitment_management.asset.applicant_gazette.entity.enums.ApplyingRank;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
