+++ src/main/java/lk/recruitment_management/asset/process_management/online_applicant_registration/entity/ConformationToken.java
@@ -1,0 +1,36 @@
+package lk.recruitment_management.asset.process_management.online_applicant_registration.entity;
+
+import lombok.*;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+import java.util.UUID;
+
+@Entity
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@EqualsAndHashCode
+@ToString
+public class ConformationToken {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+
+    private String token;
+
+    @Column(unique = true, nullable = false)
+    private String email;
+
+    private LocalDateTime createDate;
+
+    private LocalDateTime endDate;
+
+    public ConformationToken(String email) {
+        this.email = email;
+        this.token = UUID.randomUUID().toString();
+        this.createDate = LocalDateTime.now();
+        this.endDate = createDate.plusDays(1);
+    }
+}
